Update deps, break up big files.
