+++ lib/floki_util.ex
@@ -10,23 +10,33 @@ defmodule Zamrazac.FlokiUtil do
   and finally wrapping the image tag in an anchor to the original image url.
   """
   def walk_dom({tag, attributes, children}, _image_storage_path), do: {tag, attributes, children}
+
   def walk_dom(elements, image_storage_path) when is_list(elements) do
     for element <- elements do
       case element do
-        text when is_binary(text) -> text
+        text when is_binary(text) ->
+          text
+
         {"img", attributes, children} ->
           attrs = attributes_to_keywords(attributes)
           image_src = attrs[:src]
           IO.inspect("Referenced image #{image_src}", limit: :infinity)
-          {dithered_file_encoded, _temp_image_path, is_local} = convert_image(image_src, image_storage_path)
-          patched_attrs = Keyword.put(attrs, :src, dithered_file_encoded) |> keywords_to_attributes()
+
+          {dithered_file_encoded, _temp_image_path, is_local} =
+            convert_image(image_src, image_storage_path)
+
+          patched_attrs =
+            Keyword.put(attrs, :src, dithered_file_encoded) |> keywords_to_attributes()
+
           if is_local do
-            {"img", patched_attrs, walk_dom(children, image_storage_path) }
+            {"img", patched_attrs, walk_dom(children, image_storage_path)}
           else
-            {"a", [{"href", image_src}],[{"img", patched_attrs, walk_dom(children, image_storage_path) }]}
+            {"a", [{"href", image_src}],
+             [{"img", patched_attrs, walk_dom(children, image_storage_path)}]}
           end
+
         {tag, attributes, children} ->
-          {tag, attributes, walk_dom(children, image_storage_path) }
+          {tag, attributes, walk_dom(children, image_storage_path)}
       end
     end
   end
@@ -36,7 +46,7 @@ defmodule Zamrazac.FlokiUtil do
   """
   def attributes_to_keywords(attributes) do
     attributes
-    |> Enum.map(fn({key, val}) ->
+    |> Enum.map(fn {key, val} ->
       {String.to_atom(key), val}
     end)
   end
@@ -46,7 +56,7 @@ defmodule Zamrazac.FlokiUtil do
   """
   def keywords_to_attributes(keywords) do
     keywords
-    |> Enum.map(fn({key, val}) ->
+    |> Enum.map(fn {key, val} ->
       {Atom.to_string(key), val}
     end)
   end
@@ -58,23 +68,29 @@ defmodule Zamrazac.FlokiUtil do
     temp_image_name = Zamrazac.Util.shahexhash(url)
     temp_image_path = Path.join(image_storage_path, temp_image_name)
     uri = URI.parse(url)
-    case uri.scheme do
-      "file" ->
+
+    cond do
+      uri.scheme in ["http", "https"] ->
         temp_dithered_image_path = "#{temp_image_path}_dithered.png"
-        ^temp_image_path = maybe_download_image(temp_image_path, Path.expand(Path.join(uri.host, uri.path)), true)
-        {dithered_file_encoded, ^temp_dithered_image_path} = maybe_dither_image(temp_dithered_image_path, temp_image_path)
+        ^temp_image_path = maybe_download_image(temp_image_path, url, false)
+
+        {dithered_file_encoded, ^temp_dithered_image_path} =
+          maybe_dither_image(temp_dithered_image_path, temp_image_path)
+
+        {dithered_file_encoded, temp_image_path, false}
+
+      uri.scheme == "file" ->
+        temp_dithered_image_path = "#{temp_image_path}_dithered.png"
+
+        ^temp_image_path =
+          maybe_download_image(temp_image_path, Path.expand(Path.join(uri.host, uri.path)), true)
+
+        {dithered_file_encoded, ^temp_dithered_image_path} =
+          maybe_dither_image(temp_dithered_image_path, temp_image_path)
+
         {dithered_file_encoded, temp_image_path, true}
-      "http" ->
-        temp_dithered_image_path = "#{temp_image_path}_dithered.png"
-        ^temp_image_path = maybe_download_image(temp_image_path, url, false)
-        {dithered_file_encoded, ^temp_dithered_image_path} = maybe_dither_image(temp_dithered_image_path, temp_image_path)
-        {dithered_file_encoded, temp_image_path, false}
-      "https" ->
-        temp_dithered_image_path = "#{temp_image_path}_dithered.png"
-        ^temp_image_path = maybe_download_image(temp_image_path, url, false)
-        {dithered_file_encoded, ^temp_dithered_image_path} = maybe_dither_image(temp_dithered_image_path, temp_image_path)
-        {dithered_file_encoded, temp_image_path, false}
-      _ ->
+
+      true ->
         IO.inspect("\tFailed to locate image at  #{url}...", limit: :infinity)
         {"", "", true}
     end
@@ -85,17 +101,18 @@ defmodule Zamrazac.FlokiUtil do
   """
   def maybe_download_image(image_path, url, is_local) do
     case File.exists?(image_path) do
-     true ->
+      true ->
         IO.inspect("\tReusing image #{image_path}...", limit: :infinity)
         image_path
-     false ->
+
+      false ->
         if is_local do
           IO.inspect("\tCopying local image from #{url} to #{image_path}...", limit: :infinity)
           System.cmd("cp", [url, image_path])
           image_path
         else
           IO.inspect("\tDownloading image #{image_path}...", limit: :infinity)
-          System.cmd("curl", [url, "-L", "-o", image_path] )
+          System.cmd("curl", [url, "-L", "-o", image_path])
           image_path
         end
     end
@@ -108,11 +125,21 @@ defmodule Zamrazac.FlokiUtil do
     case File.exists?(image_path) do
       true ->
         IO.inspect("\tReusing dithered image #{image_path}...", limit: :infinity)
+
       false ->
         IO.inspect("\tConverting dithered image #{image_path}...", limit: :infinity)
-        System.cmd("convert", [source_image_path, "-colorspace", "Gray", "-ordered-dither", "8x8", image_path])
-     end
-     dithered_file_encoded = Zamrazac.Util.get_file_as_data_uri(image_path, "image/png")
-      {dithered_file_encoded, image_path}
+
+        System.cmd("convert", [
+          source_image_path,
+          "-colorspace",
+          "Gray",
+          "-ordered-dither",
+          "8x8",
+          image_path
+        ])
+    end
+
+    dithered_file_encoded = Zamrazac.Util.get_file_as_data_uri(image_path, "image/png")
+    {dithered_file_encoded, image_path}
   end
 end
